refactor(training): domain/entities decomposto por agregado (Fase 5.3)
Substitui `src/training/domain/entities.py` (monolítico, 626 linhas, 13
dataclasses) por pacote `domain/entities/` com 8 submódulos por agregado
e `__init__.py` que re-exporta toda a surface pública.

Submódulos:
- sessions.py       → TrainingSession + SessionObjective
- blocks.py         → SessionBlock
- wellness.py       → WellnessPre + WellnessPost
- attendance.py     → AttendanceRecord
- execution.py      → ExecutionRecord
- communication.py  → FeedbackThread + AttentionQueueItem + Recommendation
- eligibility.py    → AthleteIneligibilityDeclaration
- planning.py       → Mesocycle + Microcycle

Zero-gap:
- `training.domain.entities.<Nome>` continua resolvendo (package `__init__`
  re-exporta as 13 dataclasses + facade dos 12 StrEnums via
  `..common.enums`). Policies, api/, infrastructure/ e testes não mudam.
- test_public_surface.py e test_route_inventory.py: PASS.
- Suite training: 375 passed, 19 skipped (idêntico a Fase 5.2).
- test_exception_taxonomy.py: 47/47 PASS.

Motivação: módulo `entities.py` atingiu 23KB/626 linhas, acumulando
agregados distintos (sessão, bloco, wellness, comunicação, planejamento)
em um único arquivo. A decomposição por agregado casa com o recorte do
pacote `application/` (Fase 3) e prepara o split equivalente em
`infrastructure/repository.py` (Fase 5.4).
